// techsapphire/types.ts
export interface SymbolDefinition {
  symbol: string;
  type: 'function' | 'class' | 'const' | 'let' | 'var' | 'id' | 'animation';
  line: number;
  algorithm?: string;
}

export interface CrossReference {
  // File where the symbol is USED
  sourceFile: string;
  // File where the symbol is DEFINED
  targetFile: string;
  symbol: string;
  // line in the sourceFile where the symbol is used
  line: number;
}

export interface DeepFileAnalysis {
  fileName: string;
  // Symbols this file DEFINES
  definitions: SymbolDefinition[];
  // Symbols from OTHER files that this file USES
  usages: CrossReference[];
}